+++ game.h
@@ -1,10 +1,13 @@
 #ifndef GAME_H
 #define GAME_H
 
+#include <SDL_types.h>
+
 /* return values of towergame */
 #define GAME_FINISHED 0
 #define GAME_DIED     1
 #define GAME_ABBORTED 2
+
 
 /* load data */
 void gam_init(void);
@@ -16,7 +19,7 @@ void gam_done(void);
 void gam_newgame(void);
 
 /* initializes the tower specific data structures */
-void gam_loadtower(int tow);
+void gam_loadtower(Uint8 tow);
 
 /* leave toppler at the base of the tower */
 void gam_arrival(void);
@@ -27,9 +30,9 @@ void gam_arrival(void);
  GAME_FINISHED if the tower has been finished successfully
  GAME_DIED     if the toppler died
  GAME_ABBORTED if the game was abborted (ESC)*/
-int gam_towergame(int &anglepos, int &resttime);
+int gam_towergame(Uint8 &anglepos, Uint16 &resttime);
 
 /* pick up the toppler at the base of the tower */
-void gam_pick_up(int anglepos, int time);
+void gam_pick_up(Uint8 anglepos, Uint16 time);
 
 #endif
